requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
python-dotenv
streamlit